remove the page on extensions
The docs are primarily for app developers, but extensions are
primarily being used for runtimes, and that's what all the
examples were for.

This commit adds a quick description of extensions to the intro,
in order to explain .locale and .debug extensions (which devs will
encounter). The page itself has been moved to the wiki at:

https://github.com/flatpak/flatpak/wiki/Extensions