Try this a different way
Use a combination of :doc: and :raw: to get the cross-reference
working. This commit also adds the command reference to the main
toc. I have to use a http link there instead of a reference to
the document, since sphinx complains about command-reference.rst
not having a title even if I override the title in the toc :-(