Restructure and update section on building
Break the section into separate pages, particularly because the
page on manifests is likely to be long.

This change moves some content from the old "working with the
sandbox" page into other pages, and then removes the remaining
content. Some of this should probably come back as part of a
"developing with flatpak" section.