Add explanation of build options
Alternatively, it may be good to drop the build options from the guide altogether and document them somewhere else.

npm has nine creatively named log levels that can be used to control which errors get logged, but an electron developer would already be familiar with this.

Maybe the important thing here is to inform the user that they can set build-time environment variables in the build options, however, this isn't electron specific. It may make more sense to explain this in a more general build system guide.