Migrate html_context.css_files to html_css_files
The use of html_context = {'css_files': []} replaces stylesheets
provided by default (pygments.css and css/theme.css), which
causes the generated documentation to be unformatted due to
incomplete CSS specification.

There's recomendation of using app.add_stylesheet('file.css'),
but it is deprecated since 1.8.0, was going to removed in
Sphinx 4.0 and then it was postponed to Sphinx 6.0.

Now, html_css_files = [] was added in 1.8.0 to append CSS
files, and will solve #361 by including flatpak-docs stylesheets
to the default CSS stylesheets.