Makefile: Use sphinx-build as the default sphinx command
And fallback to `sphinx-build-3` only if it does not exist.

The default build instructions now use a python virtual environment and
inside that pip installs `sphinx-build`. Since the `<venv>/bin`
is exported as the first item in `$PATH` once activated, we can ensure
that `sphinx-build` from venv is used first.

Outside the venv if `sphinx-build` is not found `sphinx-build-3`
is tried.

Closes https://github.com/flatpak/flatpak-docs/issues/477